Extract shared boundary diff plumbing; add pilot re-engagement kit
- New core/boundary_diff.py (351 lines): the semantics-free unified-diff
  plumbing shared by the boundary check family — DiffHunk/DiffFile/
  ResolvedFileText, parse_unified_diff, hunk application, and text
  resolution. codex_boundary.py shrinks 1,526 → 1,218 lines and
  re-exports the moved names so the pre-split import surface is
  unchanged; host_boundary now imports the plumbing from the shared
  module. First step of the codex_boundary split flagged in the
  strategic review; no behavior change (diagnostic output byte-equal).
- docs/design-partner-reengagement.md: the IE-fix follow-up message
  template for stalled pilots, the audit --host first-touch for new
  prospects, and the three metrics to capture from a re-run.
- marketing/demo-agent-weakens-gate.sh: executable driver for recording
  the 90-second gate-removal demo (asciinema) per the existing script.
